% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_to_function.R
\name{bind_to_function}
\alias{bind_to_function}
\title{Enclose arguments in a function environment}
\usage{
bind_to_function(f, ...)
}
\arguments{
\item{f}{The function to which arguments are bound.}

\item{...}{Named arguments to bind into the closure environment.}
}
\value{
A function with a single formal argument \code{param}.
}
\description{
Takes a function \code{f} and named arguments, and returns a closure of
\code{f} that depends on a single argument \code{param}. Used internally when
building custom move closures in \code{bind_moves}.
}
\references{
Initial code comes from the \code{partially_apply} function in
  the \code{rodeint} package (\code{richfitz/rodeint}).
}
\seealso{
\code{\link{custom_moves}}.
}
\author{
Initial code by Rich FitzJohn (see references), adapted by Thibaut
  Jombart.
}
