% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_param}
\alias{create_param}
\alias{outbreaker_store}
\title{Initianilizes outputs for outbreaker}
\usage{
create_param(data = outbreaker_data(), config = create_config())
}
\arguments{
\item{data}{A list of data items as returned by \code{outbreaker_data}, or
arguments passed to this function.}

\item{config}{A list of settings as returned by \code{create_config}, or
arguments passed to this function.}
}
\value{
A list containing two components \code{$store} and
\code{$current}. \code{store} is a list with the class
\code{outbreaker_store}, used for storing 'saved' states of the
MCMC. \code{current} is a list with the class \code{outbreaker_param}, used
for storing 'current' states of the MCMC. \cr \cr

\code{outbreaker_store} class content:
\itemize{

 \item \code{size}: The length of the list, corresponding to the number of
samples saved from the MCMC.

 \item \code{step}: A vector of integers of length \code{size}, storing the
steps of the MCMC corresponding to the saved samples.

 \item \code{post}: A numeric vector of length \code{size}, storing
log-posterior values.

 \item \code{like}: A numeric vector of length \code{size}, storing
log-likelihood values.

 \item \code{prior}: A numeric vector of length \code{size},
storing log-prior values.

 \item \code{alpha}: A list of length \code{size}. Each item of the list is
an integer vector of length \code{data$N}, storing indices (from 1 to N) of
infectors for each case.

 \item \code{t_inf}: A list of length \code{size}. Each item of the list is
an integer vector of length \code{data$N}, storing dates of infections for
each case.

 \item \code{mu}: A numeric vector of length \code{size}, storing values of
the mutation rate.

 \item \code{kappa}: A list of length \code{size}. Each item of the list is
an integer vector of length \code{data$N}, storing the number of generations
before the last sampled ancestor for each case.

 \item \code{pi}: A numeric vector of length \code{size}, storing values of
the reporting probability.

 \item \code{eps}: Finlay - complete

 \item \code{lambda}: Finlay - complete

 \item \code{counter}: A counter used to keep track of the current iteration
of the MCMC (used internally).

}


\code{outbreaker_param} class content:
\itemize{

 \item \code{alpha}: An integer vector of length \code{data$N}, storing
indices (from 1 to N) of infectors for each case.

 \item \code{t_inf}: An integer vector of length \code{data$N}, storing dates
of infections for each case.

 \item \code{mu}: The value of the mutation rate.

 \item \code{kappa}: An integer vector of length \code{data$N}, storing the
number of generations before the last sampled ancestor for each case.

 \item \code{pi}: The value of the reporting probability.

 \item \code{eps}: Finlay - complete

 \item \code{lambda}: Finlay - complete

}
}
\description{
This function creates initial outputs and parameter states for outbreaker.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
