% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_outbreak}
\alias{fake_outbreak}
\title{Small simulated outbreak}
\format{An object of class \code{list} of length 4.}
\usage{
fake_outbreak
}
\description{
This dataset is a small simulated outbreak used to illustrate outbreaker.
It was generated using \code{simOutbreak}.
}
\author{
Thibaut Jombart \email{t_jombart@imperial_ac_uk}
}
\keyword{datasets}

