% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.rand.R
\name{fast.log.runif}
\alias{fast.log.rnorm}
\alias{fast.log.rnorm1}
\alias{fast.log.runif}
\alias{fast.log.runif1}
\alias{fast.rnorm}
\alias{fast.rnorm1}
\alias{fast.runif}
\title{Fast random number generation}
\usage{
fast.log.runif(size.batch = 50000)

fast.log.runif1(size.batch = 50000)

fast.rnorm(mean = 0, sd = 1, size.batch = 50000)

fast.rnorm1(mean = 0, sd = 1, size.batch = 50000)
}
\arguments{
\item{size.batch}{the size of the pre-generated vectors of values; larger batches lead to faster computations but require more RAM.}
}
\description{
These functions use closure programming for fast generation of random numbers from various distributions.
\itemize{
\item \code{fast.log.runif} creates a function which generates 'n' (logged) values from Unif(0,1)
\item \code{fast.runif1} creates an optimized function generating a single (logged) value from Unif(0,1)
\item \code{fast.rnorm} creates a function which generates 'n' values from Norm(mean, sd)
\item \code{fast.rnorm1} creates an optimized function generating a single value from Norm(mean, sd)
}
}
\examples{
x <- fast.log.runif(10)
environment(x)$values
x(6)
environment(x)$counter
x(4)
environment(x)$counter
x(3)
environment(x)$counter
environment(x)$values

}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

