% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.rand.vec.R
\name{fast.log.runif}
\alias{fast.log.runif}
\alias{fast.log.runif1}
\title{Fast random uniform number generation}
\usage{
fast.log.runif(size.batch = 50000)

fast.log.runif1(size.batch = 50000)
}
\arguments{
\item{size.batch}{the size of the pre-generated vectors of values; larger batches lead to faster computations but require more RAM.}
}
\description{
These functions use closure programming for fast generation of (logged) random numbers from a uniform distribution on [0;1].
\itemize{
\item \code{fast.log.runif} creates a function which generates 'n' values
\item \code{fast.runif1} creates an optimized function generating a single value
}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

