% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accuracy.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Compute accuracy of outbreak reconstruction}
\usage{
get_accuracy(out, true_tree, burnin = 0)
}
\arguments{
\item{out}{An object of class \code{outbreaker_chains}.}

\item{true_tree}{A data frame with \code{from} and \code{to} columns
representing the true transmission tree.}

\item{burnin}{The number of iterations to be discarded as burnin.}
}
\value{
A numeric vector of accuracy values (between 0 and 1), one per
  posterior sample.
}
\description{
Computes the proportion of correctly assigned infectors in each posterior
tree compared to a known true tree. Only useful in simulation studies.
}
\examples{
\dontrun{
data(fake_outbreak)
out <- outbreaker(
  data = outbreaker_data(
    dates = fake_outbreak$onset,
    dna = fake_outbreak$dna,
    w_dens = fake_outbreak$w
  ),
  config = create_config(n_iter = 500, sample_every = 50)
)
true_tree <- data.frame(
  from = as.character(fake_outbreak$ances),
  to = as.character(seq_along(fake_outbreak$onset))
)
get_accuracy(out, true_tree)
}

}
\seealso{
\code{\link{get_trees}}, \code{\link{outbreaker_chains}}.
}
\author{
Cyril Geismar (\email{c.geismar21@imperial.ac.uk}).
}
