% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trees.R
\name{get_trees}
\alias{get_trees}
\title{Extract posterior transmission trees}
\usage{
get_trees(out, burnin = 0, kappa = FALSE, t_inf = FALSE, ...)
}
\arguments{
\item{out}{An object of class \code{outbreaker_chains}.}

\item{burnin}{The number of iterations to be discarded as burnin.}

\item{kappa}{Logical. If \code{TRUE}, includes \code{kappa} values.}

\item{t_inf}{Logical. If \code{TRUE}, includes infection times.}

\item{...}{Additional vectors to include as columns. Each vector must have
the same length as the number of cases. The argument name is used as a
column suffix (e.g. \code{date = onset_dates} creates \code{from_date}
and \code{to_date}).}
}
\value{
A list of data frames, one per posterior sample.
}
\description{
Extracts a list of transmission trees from an \code{outbreaker_chains}
object, one per posterior sample. Each tree is a data frame with \code{from}
and \code{to} columns, optionally augmented with \code{kappa}, \code{t_inf},
and user-supplied columns.
}
\examples{
\dontrun{
data(fake_outbreak)
out <- outbreaker(
  data = outbreaker_data(
    dates = fake_outbreak$onset,
    dna = fake_outbreak$dna,
    w_dens = fake_outbreak$w
  ),
  config = create_config(n_iter = 500, sample_every = 50)
)
trees <- get_trees(out, date = fake_outbreak$onset)
str(trees[[1]])
}

}
\seealso{
\code{\link{get_accuracy}}, \code{\link{get_Ri}},
  \code{\link{get_offspring}}, \code{\link{outbreaker_chains}}.
}
\author{
Cyril Geismar (\email{c.geismar21@imperial.ac.uk}).
}
