% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{likelihoods}
\alias{likelihoods}
\alias{ll.all}
\alias{ll.genetic}
\alias{ll.timing}
\alias{ll.timing.infections}
\alias{ll.timing.sampling}
\title{Internal functions for outbreaker2}
\usage{
ll.timing.infections(data, param)

ll.timing.sampling(data, param)

ll.timing(data, param)

ll.genetic(data, param)

ll.all(data, param)
}
\arguments{
\item{data}{a list of named items containing input data as returned by \code{\link{outbreaker.data}}}

\item{param}{a list containing parameters as returned by \code{outbreaker.mcmc.init}}
}
\description{
These functions compute various likelihoods used in outbreaker.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

