% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.fast.rand.R
\name{make.fast.rand}
\alias{make.fast.rand}
\alias{make.fast.rand1}
\title{Fast random number generation}
\usage{
make.fast.rand(..., f = runif, batch.size = 50000, log = FALSE)

make.fast.rand1(..., f = runif, batch.size = 50000, log = FALSE)
}
\arguments{
\item{...}{arguments to be passed to a function implementing random number generation, e.g. \code{runif} or \code{rnorm}.}

\item{f}{a function implementing random number generation, e.g. \code{runif} or \code{rnorm}.}

\item{batch.size}{the size of the pre-generated vectors of values; larger batches lead to faster computations but require more RAM.}

\item{log}{a logical indicating whether generated values should be logged; defaults to FALSE.}
}
\description{
These functions use pre-generation of large random vectors to speed up random number generation.
\itemize{
\item \code{make.fast.rand} creates a function which generates 'n' values from a given distribution \code{f}.
\item \code{make.fast.runif1} creates an optimized function generating a single value from a given distribution \code{f}.
}
}
\examples{
## from Unif(0,1)
x <- make.fast.rand()
environment(x)$values
x(6)
environment(x)$counter
x(4)
environment(x)$counter
x(3)
environment(x)$counter
environment(x)$values

## check with microbenchmark
if(require(microbenchmark)){
x1 <- make.fast.rand1()
plot(microbenchmark(x1(), runif(1), times=1e3L), ylim=c(1000,3000))
}

}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}, Rich Fitzjohn
}

