% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{move.mu}
\alias{move.ances}
\alias{move.mu}
\alias{move.swap.ances}
\alias{move.t.inf}
\alias{rances}
\title{Movements of augmented data and parameters for outbreaker2}
\usage{
move.mu(data, chain, rand)

move.t.inf(data, chain, rand)

move.ances(data, chain, config, rand)

move.swap.ances(data, chain, config, rand)

rances(t.inf)
}
\arguments{
\item{data}{a list of data items as returned by \code{outbreaker.data}}

\item{chain}{a list of output items as returned by \code{outbreaker.mcmc.init}}

\item{config}{a list of settings as returned by \code{outbreaker.config}}

\item{t.inf}{a vector of infection dates}

\item{r.new}{a deviation from the previous mu}

\item{r.acc}{a logged random variate from Unif(0,1)}
}
\description{
These functions are used to move in the parameter space in the MCMC.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

