% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.chains.methods.R
\name{print.outbreaker.chains}
\alias{outbreaker.chains}
\alias{plot.outbreaker.chains}
\alias{print.outbreaker.chains}
\title{Basic methods for processing outbreaker results}
\usage{
\method{print}{outbreaker.chains}(x, n.row = 3, n.col = 8, ...)

\method{plot}{outbreaker.chains}(x, what = "post", type = c("series",
  "density"), burnin = 0, dens.all = TRUE, col = funky(x$n.runs),
  main = what, legend = TRUE, posi = "bottomleft", ...)
}
\arguments{
\item{x}{an \code{outbreaker.chains} object as returned by \code{outbreaker}.}

\item{n.row}{the number of rows to display in head and tail; defaults to 3.}

\item{n.col}{the number of columns to display; defaults to 8.}

\item{what}{a character string indicating which result to plot}

\item{type}{a character string indicating the kind of plot to be used: 'series' for time series, 'density' for a kernel density estimation}

\item{burnin}{the number of iterations to be discarded as burnin}

\item{dens.all}{a logical indicating if the overal density computed over all runs should be displayed; defaults to TRUE}

\item{col}{the colors to be used for different runs}

\item{main}{the title of the plot}

\item{legend}{a logical indicating if a legend should be added to the plot; defaults to TRUE}

\item{posi}{a character string indicating the position of the legend; defaults to "bottomleft"}
}
\description{
Several methods are defined for instances of the class \code{outbreaker.chains}, returned by \code{\link{outbreaker}}, including: \code{print}, \code{plot}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

