% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.config.methods.R
\name{print.outbreaker.config}
\alias{print.outbreaker.config}
\title{Basic methods for processing outbreaker config}
\usage{
\method{print}{outbreaker.config}(x, ...)
}
\arguments{
\item{x}{an \code{outbreaker.config} object as returned by \code{outbreaker.config}.}

\item{...}{further arguments to be passed to other methods}
}
\description{
Several methods are defined for instances of the class \code{outbreaker.config}, returned by \code{\link{outbreaker.config}}, including: \code{print}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

