% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{outbreaker.config}
\alias{outbreaker.config}
\title{#' Set and check parameter settings of outbreaker}
\usage{
outbreaker.config(..., data = NULL, config = NULL)
}
\arguments{
\item{data}{an optional list of data items as returned by \code{outbreaker.data}; if provided, this allows for further checks of the outbreaker settings.}

\item{config}{a previous set of settings, as returned by \code{outbreaker.config}}
}
\description{
This function defines settings for outbreaker.
It takes a list of named items as input, performs various checks, set defaults where arguments are missing, and return a correct list of settings. If no input is given, it returns the default settings.
}
\details{
Acceptables arguments for ... are:
\describe{

\item{n.iter}{an integer indicating the number of iterations in the MCMC,
including the burnin period}

\item{init.mu}{initial values for the mutation rates}

\item{move.ances}{a vector of logicals indicating, for each case, if the ancestry should be estimated ('moved' in the MCMC), or not, defaulting to TRUE; the vector is recycled if needed.}

\item{move.t.inf}{a vector of logicals indicating, for each case, if the dates of infection should be estimated ('moved' in the MCMC), or not, defaulting to TRUE; the vector is recycled if needed.}

\item{move.mu}{a logical indicating whether the mutation rates
should be estimated ('moved' in the MCMC), or not, all defaulting to TRUE.}

\item{n.iter}{the number of iterations of the MCMC}

\item{sample.every}{the frequency at which MCMC samples are retained for the output}

\item{sd.mu}{the standard deviation for the Normal proposal for the mutation rates}
}
}
\examples{
## see default settings
outbreaker.config()

## change defaults
outbreaker.config(move.ances=FALSE, n.iter=2e5, sample.every=1000)
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}
\seealso{
outbreaker.data to check and process data for outbreaker
}

