% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.create.loglike.R
\name{outbreaker.create.loglike}
\alias{outbreaker.create.loglike}
\title{Create likelihood functions for outbreaker}
\usage{
outbreaker.create.loglike(..., loglike = NULL)
}
\arguments{
\item{...}{a named list (see details) of functions, each computing a log-likelihood component.}

\item{loglike}{a list of functions as returned by \code{outbreaker.create.loglike}}
}
\value{
a list of named functions
}
\description{
This function creates a named list of functions, each of which are compute log-likelihood components of outbreaker.
If these functions are provided as input, they will be used. Otherwise, default functions will be used.
}
\details{
User-provided functions will be checked to ensure the right arguments are present. All log-likelihood functions should have to following arguments:
\describe{
\item{data}{a list of named items containing input data as returned by \code{\link{outbreaker.data}}}
\item{param}{a list of parameters as returned by \code{outbreaker.mcmc.init}} 
}

Note that unlike \code{outbreaker.create.moves}, likelihood functions have no mandatory names. They only need to be compatible with the movement functions used in \code{outbreaker.create.moves}. See 'details' for defaults.


The likelihood functions used by default are:
\describe{
\item{timing.infections}{corresponds to the probability of delays between cases (generation time)}
\item{timing.sampling}{corresponds to the probability of delays between infection and reporting}
\item{timing}{corresponds to the probability of all delays}
\item{genetic}{corresponds to the probability of the genetic sequences given the transmission tree}
\item{all}{corresponds to the overal probability of the data}
}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

