% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.create.moves.R
\name{outbreaker.create.moves}
\alias{outbreaker.create.moves}
\title{Create the parameter space explorer}
\usage{
outbreaker.create.moves(..., moves = NULL)
}
\arguments{
\item{...}{a list of named movement functions; see details for available names.}

\item{moves}{a list of functions as returned by \code{outbreaker.create.moves}}
}
\value{
a list of named functions
}
\description{
This function creates a named list of functions, each of which moves parameters of outbreaker.
If these functions are provided as input, they will be used. Otherwise, default functions will be used.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

