% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.create.posterior.R
\name{outbreaker.create.posteriors}
\alias{outbreaker.create.posteriors}
\title{Create posterior functions for outbreaker}
\usage{
outbreaker.create.posteriors(..., posteriors = NULL)
}
\arguments{
\item{...}{a named list (see details) of functions, each computing a log-posterior component.}

\item{posteriors}{a list of functions as returned by \code{outbreaker.create.posteriors}}
}
\value{
a list of named functions
}
\description{
This function creates a named list of functions, each of which are compute log-posterior components of outbreaker.
If these functions are provided as input, they will be used. Otherwise, default functions will be used.
}
\details{
User-provided functions will be checked to ensure the right arguments are present. All log-posterior functions should have to following arguments:
\describe{
\item{data}{a list of named items containing input data as returned by \code{\link{outbreaker.data}}}
\item{param}{a list of parameters as returned by \code{\link{outbreaker.mcmc.init}}} 
}

Note that unlike \code{outbreaker.create.moves}, posterior functions have no mandatory names. They only need to be compatible with the movement functions used in \code{outbreaker.create.moves}. See 'details' for defaults.


The posterior functions used by default are:
\describe{
\item{genetic}{corresponds to the posterior distribution of the mutation rates; defaults to \code{\link{post.genetic}}}
\item{all}{corresponds to the joint posterior; defaults to \code{\link{post.all}}}
}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

