% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.create.priors.R
\name{outbreaker.create.priors}
\alias{outbreaker.create.priors}
\title{Create prior functions for outbreaker}
\usage{
outbreaker.create.priors(..., priors = NULL)
}
\arguments{
\item{...}{a named list (see details) of functions, each computing a log-prior component.}

\item{priors}{a list of functions as returned by \code{outbreaker.create.priors}}
}
\value{
a list of named functions
}
\description{
This function creates a named list of functions, each of which are compute log-prior components of outbreaker.
If these functions are provided as input, they will be used. Otherwise, default functions will be used.
}
\details{
User-provided functions will be checked to ensure the right arguments are present. All log-prior functions should have to following arguments:
\describe{
\item{param}{a list of parameters as returned by \code{outbreaker.mcmc.init}} 
}

Note that unlike \code{outbreaker.create.moves}, prior functions have no mandatory names. They only need to be compatible with the movement functions used in \code{outbreaker.create.moves}. See 'details' for defaults.


The prior functions used by default are:
\describe{
\item{mu}{corresponds to the prior of the mutation rates; defaults to \code{\link{prior.mu}}}
\item{all}{corresponds to the joint priors of all parameters; defaults to \code{\link{prior.all}}}
}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

