% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{outbreaker.data}
\alias{outbreaker.data}
\title{Checks input data for outbreaker}
\usage{
outbreaker.data(dates, w.dens, f.dens = w.dens, dna = NULL)
}
\description{
This function performs various checks on input data given to outbreaker.
}
\details{
Acceptables arguments are:
\describe{

\item{dna}{the DNA sequences in \code{DNAbin} format (see
\code{\link[ape]{read.dna}} in the ape package); this can be imported from a
fasta file (extension .fa, .fas, or .fasta) using \code{adegenet}'s function
\link[adegenet]{fasta2DNAbin}.}

\item{dates}{a vector indicating the collection dates, provided either as
integer numbers or in a usual date format such as \code{Date} or
\code{POSIXct} format. By convention, zero will indicate the oldest date.}

\item{w.dens}{a vector of numeric values indicating the generation time
distribution, reflecting the infectious potential of a case t=0, 1, 2, ...
time steps after infection. By convention, w.dens[1]=0, meaning that an
newly infected patient cannot be instantaneously infectious. If not
standardized, this distribution is rescaled to sum to 1.}

\item{f.dens}{similar to \code{w.dens}, except that this is the distribution
of the colonization time, i.e. time interval during which the pathogen can
be sampled from the patient.}

\item{init.tree}{the tree used to initialize the MCMC. Can be either a
character string indicating how this tree should be computed, or a vector of
integers corresponding to the tree itself, where the i-th value corresponds
to the index of the ancestor of 'i' (i.e., \code{init.tree[i]} is the
ancestor of case \code{i}). Accepted character strings are "seqTrack" (uses
seqTrack output as initialize tree), "random" (ancestor randomly selected
from preceding cases), and "star" (all cases coalesce to the first case).
Note that for SeqTrack, all cases should have been sequenced.}

}
}
\examples{
x <- fakeOutbreak
outbreaker.data(dates=x$collecDates, dna=x$dat$dna, w.dens=x$w)
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}

