% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.init.mcmc.R
\name{outbreaker.init.mcmc}
\alias{outbreaker.init.mcmc}
\title{Initianilizes outputs for outbreaker}
\usage{
outbreaker.init.mcmc(param.current, param.store, loglike, priors)
}
\arguments{
\item{loglike}{a list of loglikelihood functions with enclosed data as returned by \code{create.loglike}}

\item{priors}{a list of prior functions with enclosed parameters as returned by \code{create.priors}}

\item{param}{a list of data items as returned by \code{outbreaker.create.mcmc}}
}
\description{
This function creates initial outputs and parameter states for outbreaker.
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}

