% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.mcmc.init.R
\name{outbreaker.mcmc.init}
\alias{outbreaker.mcmc.init}
\title{Initianilizes outputs for outbreaker}
\usage{
outbreaker.mcmc.init(data, config)
}
\arguments{
\item{data}{a list of data items as returned by \code{outbreaker.data}}

\item{config}{a list of settings as returned by \code{outbreaker.config}}
}
\description{
This function creates initial outputs and parameter states for outbreaker.
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}

