% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.mcmc.shape.R
\name{outbreaker.mcmc.shape}
\alias{outbreaker.mcmc.shape}
\title{Shape MCMC samples into outputs for outbreaker}
\usage{
outbreaker.mcmc.shape(chain, data)
}
\arguments{
\item{chain}{a list of output items as returned by \code{outbreaker.mcmc.init}}
}
\description{
This function shapes MCMC samples for outbreaker into a mcmc object from the package \code{coda}.
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}

