% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.mcmc.store.R
\name{outbreaker.mcmc.store}
\alias{outbreaker.mcmc.store}
\title{Stores MCMC samples for outbreaker}
\usage{
outbreaker.mcmc.store(param.current, param.store, densities, step)
}
\arguments{
\item{densities}{a list containing lists of functions computing densities, named: 'loglike' (log-likelihoods), 'priors' and 'posteriors'}

\item{step}{an integer indicating the MCMC iteration being stored}

\item{param}{a list of output items as returned by \code{outbreaker.create.mcmc}}
}
\description{
This function creates stores MCMC samples for outbreaker: augmented data and parameter states, likelihood, priors and posterior.
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}

