% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.move.R
\name{outbreaker.move}
\alias{outbreaker.move}
\title{Movements of augmented data and parameters for outbreaker2}
\usage{
outbreaker.move(moves, data, param, config, densities)
}
\arguments{
\item{param}{a list of parameters as returned by \code{outbreaker.create.mcmc}}

\item{loglike}{a list of log-likelihood functions with enclosed data as returned by \code{create.loglike}}

\item{priors}{a list of prior functions with enclosed parameters as returned by \code{create.priors}}

\item{posteriors}{a list of posterior functions with enclosed likelihood and prior functions as
returned by \code{create.posteriors}}
}
\value{
a potentially modified list of parameters as returned by \code{outbreaker.create.mcmc}
}
\description{
This function moves all parameters for outbreaker2.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

