% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.rand.vec.R
\name{outbreaker.rand.vec}
\alias{outbreaker.rand.vec}
\title{Random number pre-generation for outbreaker2}
\usage{
outbreaker.rand.vec(config)
}
\arguments{
\item{config}{a list of settings as returned by \code{outbreaker.config}}
}
\description{
This function generates pre-drawn vectors of random numbers to be used in outbreaker2.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

