% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.config.methods.R
\name{print.outbreaker_config}
\alias{print.outbreaker_config}
\title{Basic methods for processing outbreaker config}
\usage{
\method{print}{outbreaker_config}(x, ...)
}
\arguments{
\item{x}{an \code{outbreaker_config} object as returned by \code{outbreaker_config}.}

\item{...}{further arguments to be passed to other methods}
}
\description{
Several methods are defined for instances of the class \code{outbreaker_config}, returned by \code{\link{outbreaker_config}}, including: \code{print}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail_com})
}

