% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.create.mcmc.R
\name{create_mcmc}
\alias{create_mcmc}
\title{Initianilizes outputs for outbreaker}
\usage{
create_mcmc(data, config)
}
\arguments{
\item{data}{a list of data items as returned by \code{outbreaker_data}}

\item{config}{a list of settings as returned by \code{create_config}}
}
\description{
This function creates initial outputs and parameter states for outbreaker.
}
\author{
Thibaut Jombart (\email{t_jombart@imperial_ac_uk})
}

