% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker_data.R
\name{outbreaker_data}
\alias{outbreaker_data}
\title{Process input data for outbreaker}
\usage{
outbreaker_data(..., data = list(...))
}
\arguments{
\item{...}{a list of data items to be processed (see description)}

\item{data}{optionally, an existing list of data item as returned by \code{outbreaker_data}.}
}
\description{
This function performs various checks on input data given to outbreaker.  It
takes a list of named items as input, performs various checks, set defaults
where arguments are missing, and return a correct list of data input. If no
input is given, it returns the default settings.
}
\examples{

x <- fake_outbreak
outbreaker_data(dates = x$sample, dna = x$dna, w_dens = x$w)

}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
