% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker_init_mcmc.R
\name{outbreaker_init_mcmc}
\alias{outbreaker_init_mcmc}
\title{Initialises outputs for outbreaker}
\usage{
outbreaker_init_mcmc(data, param_current, param_store, loglike, priors, config)
}
\arguments{
\item{loglike}{a list of loglikelihood functions with enclosed data
as returned by \code{custom_likelihood}}

\item{priors}{a list of prior functions with enclosed parameters as
returned by \code{custom_priors}}

\item{param}{a list of data items as returned by \code{create_param}}
}
\description{
This function creates initial outputs and parameter states for outbreaker.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
