% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker_mcmc_shape.R
\name{outbreaker_mcmc_shape}
\alias{outbreaker_mcmc_shape}
\title{Shape MCMC samples into outputs for outbreaker}
\usage{
outbreaker_mcmc_shape(param, data)
}
\arguments{
\item{param}{a list of output items as returned by \code{create_mcmc}}

\item{data}{a list of data items as returned by \code{outbreaker_data}}
}
\description{
This function shapes MCMC samples for outbreaker into a data.frame with the class \code{outbreaker_chains}.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail_com})
}

