% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker_mcmc_store.R
\name{outbreaker_mcmc_store}
\alias{outbreaker_mcmc_store}
\title{Stores MCMC samples for outbreaker}
\usage{
outbreaker_mcmc_store(param_current, param_store, data, config, densities, step)
}
\arguments{
\item{data}{a list of named items containing input data as returned by
\code{\link{outbreaker_data}}}

\item{config}{a set of settings as returned by \code{\link{create_config}}}

\item{densities}{a list containing lists of functions computing densities, named: 'loglike' (log-likelihoods), 'priors' and 'posteriors'}

\item{step}{an integer indicating the MCMC iteration being stored}

\item{param}{a list of output items as returned by \code{create_mcmc}}
}
\description{
This function creates stores MCMC samples for outbreaker: augmented data and parameter states, likelihood, priors and posterior.
}
\author{
Thibaut Jombart (\email{t_jombart@imperial_ac_uk})
}

