% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker_move.R
\name{outbreaker_move}
\alias{outbreaker_move}
\title{Movements of augmented data and parameters for outbreaker2}
\usage{
outbreaker_move(moves, data, param_current, param_store, config, likelihoods,
  priors)
}
\arguments{
\item{moves}{a list of movement functions as returned by \code{bind_moves}
(internal function)}

\item{param}{a list of parameters as returned by
\code{create_mcmc}}

\item{densities}{a list containing lists of functions computing densities,
named: 'loglike' (log-likelihoods), 'priors' and 'posteriors'}
}
\value{
a potentially modified list of parameters as returned by
    \code{create_mcmc}
}
\description{
This function moves all parameters for outbreaker2.
}
\author{
Thibaut Jombart \email{t_jombart@imperial_ac_uk}
}

