% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_outbreaker_param.R
\docType{data}
\name{outbreaker_param}
\alias{outbreaker_param}
\title{Parameters and augmented data using in outbreaker2}
\format{An object of class \code{NULL} of length 0.}
\usage{
outbreaker_param
}
\description{
TBC...
}
\keyword{datasets}

