% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{chains_pal}
\alias{chains_pal}
\alias{outbreaker_palettes}
\alias{cases_pal}
\title{Color palettes used in outbreaker}
\usage{
chains_pal(n)

cases_pal(n)
}
\arguments{
\item{n}{a number of colors to be created}
}
\description{
These functions are different color palettes (color-generating functions)
used in outbreaker.
}
\examples{
plot(1:8, col = chains_pal(8), cex = 10, pch = 20)

}
\author{
Thibaut Jombart (\email{thibautjombart@gmail_com})
}
