% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{post.genetic}
\alias{post.all}
\alias{post.genetic}
\alias{post.reporting}
\title{Posteriors of outbreaker2}
\usage{
post.genetic(data, param)

post.reporting(data, param)

post.all(data, param)
}
\arguments{
\item{data}{a list of named items containing input data as returned by \code{\link{outbreaker.data}}}

\item{param}{a list containing parameters as returned by \code{outbreaker.mcmc.init}}
}
\description{
These functions compute various posteriors used in outbreaker.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

