% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCTD.R
\name{simCTD}
\alias{simCTD}
\title{Simulate contact tracing data from a simOutbreak object}
\usage{
simCTD(outbreak, eps = 1, xi = 0, plot = FALSE)
}
\arguments{
\item{outbreak}{a simOutbreak object}

\item{eps}{the reporting probability of infectious contacts}

\item{xi}{the scaling factor relative to eps, describing the probability of reporting contact between non-infectiously related pairs}

\item{plot}{a logical indicating if the contact tracing data should be plotted}
}
\description{
This function takes a simOutbreak object and returns a dataframe of contact pairs. The probabilites of reporting infectious and non-infectious contacts are provided as parameters.
}
\author{
Finlay Campbell (\email{f.campbell15@imperial.ac.uk})
}
