% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated outbreak dataset}
\alias{fakeOutbreak}
\alias{simulated outbreak dataset}
\title{Toy outbreak dataset used to illustrate outbreaker}
\description{
This toy outbreak dataset was simulated using \code{simOutbreak}.
This dataset is a list containing the following components: \itemize{ \item
dat: the data, output of simOutbreak; see \code{dat$call} for the actual
command line that was used.  \item w: the generation time distribution.
\item collecDates: simulated collection dates dates.  \item res: the results
of \code{outbreaker.parallel}; see \code{res$call} for the actual command
line that was used.  }
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}
\keyword{datasets}

