% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{fake_outbreak}
\alias{fake_outbreak}
\title{Small simulated outbreak}
\format{
An object of class \code{list} of length 6.
}
\usage{
fake_outbreak
}
\description{
This dataset is a small (30 cases) simulated outbreak originally used to
illustrate \code{outbreaker}, and used for the same purposes in
\code{outbreaker2}.  This list contains the following:
}
\details{
\itemize{

\item \code{$onset}: A vector of integers representing dates of onset.

\item \code{$sample}: A vector of integers representing the dates of isolation.

\item \code{$dna}: A DNAbin matrix of pathogen genome sequences.

\item \code{$w}: A numeric vector giving the empirical distribution of the
generation time.

\item \code{$ances}: A vector of integers indicating, for each case, the true
infectors. \code{NA} represents imported cases.

}
}
\examples{
names(fake_outbreak)
fake_outbreak

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\keyword{datasets}
