% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.rand.R
\name{make.fast.log.runif}
\alias{make.fast.log.rnorm}
\alias{make.fast.log.rnorm1}
\alias{make.fast.log.runif}
\alias{make.fast.log.runif1}
\alias{make.fast.runif}
\title{Fast random number generation}
\usage{
make.fast.log.runif(batch.size = 50000)
}
\arguments{
\item{batch.size}{the size of the pre-generated vectors of values; larger batches lead to faster computations but require more RAM.}
}
\description{
These functions use closure programming for fast generation of random numbers from various distributions.
\itemize{
\item \code{make.fast.log.runif} creates a function which generates 'n' (logged) values from Unif(0,1)
\item \code{make.fast.runif1} creates an optimized function generating a single (logged) value from Unif(0,1)
\item \code{make.fast.rnorm} creates a function which generates 'n' values from Norm(mean, sd)
\item \code{make.fast.rnorm1} creates an optimized function generating a single value from Norm(mean, sd)
}
}
\examples{
x <- make.fast.log.runif(10)
environment(x)$values
x(6)
environment(x)$counter
x(4)
environment(x)$counter
x(3)
environment(x)$counter
environment(x)$values

}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

