% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{modify_defaults}
\alias{modify_defaults}
\alias{ralpha}
\title{Internal functions for outbreaker2}
\usage{
modify_defaults(defaults, x, strict = TRUE)

ralpha(t_inf)
}
\arguments{
\item{defaults}{a list containing default arguments}

\item{x}{in \code{modify_defaults}, a list with user-provided arguments; in
\code{add_to_context}, a function or an environment.}

\item{strict}{a logical indicating if errors shoul be returned when 'x'
contains items not in 'defaults'}

\item{t_inf}{a vector of infection dates}
}
\description{
These functions are meant for internal use in outbreaker2. They are not exported, and their API might change in the future.
}
\details{
\describe{
\item{modify_default}{modify default arguments using user-provided values}
}
}
\author{
Rich Fitzjohn, Thibaut Jombart
}

