% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{ll.timing.infections}
\alias{ll.genetic}
\alias{ll.timing.infections}
\title{Internal functions for outbreaker2}
\usage{
ll.timing.infections(times, ances, log.w)

ll.genetic(D, ances, mu, gen.length)
}
\arguments{
\item{times}{a vector of integers indicating dates of infections of the cases}

\item{ances}{a vector of indices of ancestors}

\item{log.w}{a vector of log probabilities of time intervals, starting at p(T=0)}

\item{D}{a matrix of pairwise genetic distances}

\item{mu}{a mutation rate}
}
\description{
These functions represent the internal machinary behind outbreaker2: priors, likelihood functions, movements of parameters and augmented data.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

