% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{ll.timing.infections}
\alias{ll.all}
\alias{ll.genetic}
\alias{ll.timing.infections}
\alias{ll.timing.sampling}
\title{Internal functions for outbreaker2}
\usage{
ll.timing.infections(times, ances, log.w)

ll.timing.sampling(times, sampling.times, log.f)

ll.genetic(D, ances, mu, gen.length)

ll.all(times, ances, log.w, D, mu, gen.length)
}
\arguments{
\item{times}{a vector of integers indicating dates of infections of the cases}

\item{ances}{a vector of indices of ancestors}

\item{log.w}{a vector of log probabilities of time intervals (between infections), starting at p(T=0)}

\item{log.f}{a vector of log probabilities of time intervals (from infection to collection), starting at p(T=0)}

\item{D}{a matrix of pairwise genetic distances}

\item{mu}{a mutation rate}

\item{gen.length}{length of the genetic sequences}

\item{sampling.time}{a vector of integers indicating dates of sampling/observation/reporting of the cases}
}
\description{
These functions represent the internal machinary behind outbreaker2: priors, likelihood functions, movements of parameters and augmented data.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

