% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{modify.defaults}
\alias{add.to.context}
\alias{modify.defaults}
\title{Internal functions for outbreaker2}
\usage{
modify.defaults(defaults, x, strict = TRUE)

add.to.context(x, objects)
}
\arguments{
\item{defaults}{a list containing default arguments}

\item{x}{in \code{modify.defaults}, a list with user-provided arguments; in \code{add.to.context}, a function.}

\item{strict}{a logical indicating if errors shoul be returned when 'x' contains items not in 'defaults'}

\item{objects}{a list of objects to be added to the environment of 'x'; all items need to be named.}
}
\description{
These functions are meant for internal use in outbreaker2. They are not exported, and their API might change in the future.
}
\details{
\describe{
\item{modify.default}{modify default arguments using user-provided values}
\item{add.to.function.context}{add a set of objects to the environment of a function}
}
}
\author{
Rich Fitzjohn, Thibaut Jombart
}

