% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{likelihoods}
\alias{likelihoods}
\alias{ll.all}
\alias{ll.genetic}
\alias{ll.reporting}
\alias{ll.timing}
\alias{ll.timing.infections}
\alias{ll.timing.sampling}
\title{Internal functions for outbreaker2}
\usage{
ll.timing.infections(data, param, i = NULL)

ll.timing.sampling(data, param, i = NULL)

ll.timing(data, param, i = NULL)

ll.genetic(data, param, i = NULL)

ll.reporting(data, param, i = NULL)

ll.all(data, param, i = NULL)
}
\arguments{
\item{data}{a list of named items containing input data as returned by \code{\link{outbreaker.data}}}

\item{param}{a list containing parameters as returned by \code{outbreaker.mcmc.init}}

\item{i}{an optional vector of integers, indicating subset of cases included in the likelihood computation; if NULL (default), all cases are used.}
}
\description{
These functions compute various likelihoods used in outbreaker.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

