% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{ll.timing.infections}
\alias{ll.all}
\alias{ll.genetic}
\alias{ll.timing.infections}
\alias{ll.timing.sampling}
\title{Internal functions for outbreaker2}
\usage{
ll.timing.infections(times, log.w, ances)

ll.timing.sampling(times, log.f, sampling.times)

ll.genetic(D, gen.length, ances, mu)

ll.all(times, D, gen.length, log.w, ances, mu)
}
\arguments{
\item{times}{a vector of integers indicating dates of infections of the cases}

\item{log.w}{a vector of log probabilities of time intervals (between infections), starting at p(T=0)}

\item{ances}{a vector of indices of ancestors}

\item{log.f}{a vector of log probabilities of time intervals (from infection to collection), starting at p(T=0)}

\item{D}{a matrix of pairwise genetic distances}

\item{gen.length}{length of the genetic sequences}

\item{mu}{a mutation rate}

\item{sampling.time}{a vector of integers indicating dates of sampling/observation/reporting of the cases}
}
\description{
These functions compute various likelihoods used in outbreaker.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

