% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{move.mu}
\alias{move.alpha}
\alias{move.mu}
\alias{move.pi}
\alias{move.swap.alpha}
\alias{move.t.inf}
\alias{ralpha}
\title{Movements of augmented data and parameters for outbreaker2}
\usage{
move.mu(data, param, config, rand)

move.t.inf(data, param, config, rand)

move.alpha(data, param, config, rand)

move.swap.alpha(data, param, config, rand)

ralpha(t.inf)

move.pi(data, param, config, rand)
}
\arguments{
\item{data}{a list of data items as returned by \code{outbreaker.data}}

\item{param}{a list of parameters as returned by \code{outbreaker.mcmc.init}}

\item{config}{a list of settings as returned by \code{outbreaker.config}}

\item{rand}{a list of items as returned by \code{outbreaker.rand.vec}}

\item{t.inf}{a vector of infection dates}
}
\value{
a potentially modified list of parameters as returned by \code{outbreaker.mcmc.init}
}
\description{
These functions are used to move in the parameter space in the MCMC.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

