% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{move.mu}
\alias{move.ances}
\alias{move.mu}
\alias{move.t.inf}
\alias{rances}
\title{Movements of augmented data and parameters for outbreaker2}
\usage{
move.mu(data, chain, r.new, r.acc)

move.t.inf(data, chain, r.acc)

rances(t.inf)

move.ances(data, chain, r.acc)
}
\arguments{
\item{data}{a list of data items as returned by \code{outbreaker.data}}

\item{chain}{a list of output items as returned by \code{outbreaker.mcmc.init}}

\item{r.new}{a deviation from the previous mu}

\item{r.acc}{a logged random variate from Unif(0,1)}

\item{t.inf}{a vector of infection dates}
}
\description{
These functions are used to move in the parameter space in the MCMC.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

