% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moves.R
\name{move.mu}
\alias{move.ances}
\alias{move.mu}
\alias{move.t.inf}
\alias{rances}
\title{Movements of augmented data and parameters for outbreaker2}
\usage{
move.mu(D, gen.length, ances, mu, dev, lunif)

move.t.inf(t.inf, log.w, log.f, ances, sampling.times, lunif)

rances(t.inf)

move.ances(t.inf, sampling.times, D, gen.length, log.w, log.f, ances, mu, lunif)
}
\arguments{
\item{D}{a matrix of pairwise genetic distances}

\item{gen.length}{length of the genetic sequences}

\item{ances}{a vector of indices of ancestors}

\item{mu}{a mutation rate}

\item{dev}{a deviation from the previous mu}

\item{lunif}{a logged random variate from Unif(0,1)}

\item{t.inf}{a vector of integers indicating dates of infections of the cases}

\item{log.w}{a vector of log probabilities of time intervals (between infections), starting at p(T=0)}
}
\description{
These functions are used to move in the parameter space in the MCMC.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

