% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.config.R
\name{outbreaker.config}
\alias{outbreaker.config}
\title{Set and check parameter settings of outbreaker}
\usage{
outbreaker.config(..., data = NULL, config = NULL)
}
\arguments{
\item{...}{settings to be passed to outbreaker}

\item{data}{an optional list of data items as returned by \code{outbreaker.data}; if provided, this allows for further checks of the outbreaker setings.}

\item{config}{a previous set of settings, as returned by \code{outbreaker.config}}
}
\description{
This function defines settings for outbreaker.
It takes a list of named items as input, performs various checks, set defaults where arguments are missing, and return a correct list of settings. If no input is given, it returns the default settings.
}
\details{
Acceptables arguments for ... are:
\describe{
\item{init.tree}{the tree used to initialize the MCMC. Can be either a
character string indicating how this tree should be computed, or a vector of
integers corresponding to the tree itself, where the i-th value corresponds
to the index of the ancestor of 'i' (i.e., \code{init.tree[i]} is the
ancestor of case \code{i}). Accepted character strings are "seqTrack" (uses
seqTrack output as initialize tree), "random" (ancestor randomly selected
from preceding cases), and "star" (all cases coalesce to the first case).
Note that for SeqTrack, all cases should have been sequenced.}

\item{n.iter}{an integer indicating the number of iterations in the MCMC,
including the burnin period}

\item{init.mu}{initial values for the mutation rates}

\item{init.kappa}{a (recycled) vector of integers indicating the initial values of kappa; defaults to 1.}

\item{move.ances}{a vector of logicals indicating, for each case, if the ancestry should be estimated ('moved' in the MCMC), or not, defaulting to TRUE; the vector is recycled if needed.}

\item{move.t.inf}{a vector of logicals indicating, for each case, if the dates of infection should be estimated ('moved' in the MCMC), or not, defaulting to TRUE; the vector is recycled if needed.}

\item{move.mu}{a logical indicating whether the mutation rates
should be estimated ('moved' in the MCMC), or not, all defaulting to TRUE.}

\item{move.kappa}{a logical indicating whether the number of generations between two successive cases
should be estimated ('moved' in the MCMC), or not, all defaulting to TRUE.}

\item{n.iter}{the number of iterations of the MCMC}

\item{sample.every}{the frequency at which MCMC samples are retained for the output}

\item{sd.mu}{the standard deviation for the Normal proposal for the mutation rates}

\item{prop.ances.move}{the proportion of ancestries to move at each iteration of the MCMC}

\item{batch.size}{the size of the batch of random number pre-generated}

\item{paranoid}{a logical indicating if the paranoid mode should be used; this mode is used for performing additional tests during outbreaker; it makes computations substantially slower and is mostly used for debugging purposes.}

\item{min.date}{earliest infection date possible, expressed as days since the first sampling;}

\item{max.kappa}{an integer indicating the largest number of generations between any two linked cases; defaults to 5}
}
}
\examples{
## see default settings
outbreaker.config()

## change defaults
outbreaker.config(move.ances=FALSE, n.iter=2e5, sample.every=1000)


}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}
\seealso{
outbreaker.data to check and process data for outbreaker
}

