% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.create.moves.R
\name{outbreaker.create.moves}
\alias{outbreaker.create.moves}
\title{Create the parameter space explorer}
\usage{
outbreaker.create.moves(..., moves = NULL)
}
\arguments{
\item{...}{a named list (see details), each component being a list movement functions defining movements of parameters or augmented data; see details for available movements and the corresponding names.}

\item{moves}{a list of functions as returned by \code{outbreaker.create.moves}}
}
\value{
a list of named functions
}
\description{
This function creates a named list of functions, each of which moves parameters of outbreaker.
If these functions are provided as input, they will be used. Otherwise, default functions will be used.
}
\details{
User-provided functions will be checked to ensure the right arguments are present. All movement functions should have to following arguments:
\describe{
\item{data}{a list of named items containing input data as returned by \code{\link{outbreaker.data}}}
\item{config}{a set of settings as returned by \code{\link{outbreaker.config}}}
\item{param}{a list of parameters as returned by \code{outbreaker.mcmc.init}} 
\item{rand}{a list of items as returned by \code{outbreaker.rand.vec}} 
}


The movement functions which can be passed as arguments are:
\describe{
\item{mu}{a list of functions to move mutation rates; defaults to \code{move.mu}}
\item{t.inf}{a list of functions to move dates of infection; defaults to \code{move.t.inf}}
\item{ances}{a list of functions to move ancestries, i.e. the transmission tree; defaults to \code{move.ances} and \code{move.swap.ances}}
}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

