% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.create.moves.R
\name{outbreaker.create.moves}
\alias{outbreaker.create.moves}
\title{Create the parameter space explorer}
\usage{
outbreaker.create.moves(..., moves = NULL, config = outbreaker.config())
}
\arguments{
\item{...}{a named list of movement functions for parameters or augmented data; see details for available movements and the corresponding names.}

\item{moves}{a list of functions as returned by \code{outbreaker.create.moves}}

\item{config}{a list of settings as returned by \code{outbreaker.config}}
}
\value{
a list of named functions
}
\description{
This function creates a named list of functions, each of which moves parameters of outbreaker.
If these functions are provided as input, they will be used. Otherwise, default functions will be used.
}
\details{
User-provided functions will be checked to ensure the right arguments are present. All movement functions should have to following arguments:
\describe{
\item{data}{a list of named items containing input data as returned by \code{\link{outbreaker.data}}}
\item{config}{a set of settings as returned by \code{\link{outbreaker.config}}}
\item{param}{a list of parameters as returned by \code{outbreaker.mcmc.init}}
\item{rand}{a list of items as returned by \code{outbreaker.rand.vec}}
}


The movement functions which are used by default:
\describe{
\item{move.mu}{a function to move mutation rates}
\item{move.t.inf}{a function to move dates of infection}
\item{move.alpha}{a function to move ancestries, i.e. the transmission tree}
\item{move.swap.cases}{another function to move ancestries, relying on swapping ancestries (a->b becomes b->a)}
}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

