% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.mcmc.store.R
\name{outbreaker.mcmc.store}
\alias{outbreaker.mcmc.store}
\title{Stores MCMC samples for outbreaker}
\usage{
outbreaker.mcmc.store(chain, data)
}
\arguments{
\item{chain}{a list of output items as returned by \code{outbreaker.mcmc.init}}

\item{data}{a list of settings as returned by \code{outbreaker.data}}
}
\description{
This function creates stores MCMC samples for outbreaker: augmented data and parameter states, likelihood, priors and posterior.
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}

