% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.move.R
\name{outbreaker.move}
\alias{outbreaker.move}
\title{Movements of augmented data and parameters for outbreaker2}
\usage{
outbreaker.move(moves, data, config, param, rand)
}
\arguments{
\item{moves}{a set of movement functions stored in a named list as returned by \code{\link{outbreaker.create.moves}}}

\item{data}{a list of data items as returned by \code{outbreaker.data}}

\item{config}{a list of settings as returned by \code{outbreaker.config}}

\item{param}{a list of parameters as returned by \code{outbreaker.mcmc.init}}

\item{rand}{a list of items as returned by \code{outbreaker.rand.vec}}
}
\value{
a potentially modified list of parameters as returned by \code{outbreaker.mcmc.init}
}
\description{
This function moves all parameters for outbreaker2.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

