% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.R
\name{outbreaker}
\alias{outbreaker}
\title{Outbreaker2: disease outbreak reconstruction using epidemiological and genetic data}
\usage{
outbreaker(data = outbreaker.data(), config = outbreaker.config(),
  loglike = outbreaker.create.loglike(),
  priors = outbreaker.create.priors(),
  posteriors = outbreaker.create.posteriors(),
  moves = outbreaker.create.moves())
}
\arguments{
\item{data}{a list of named items containing input data as returned by \code{\link{outbreaker.data}}}

\item{config}{a set of settings as returned by \code{\link{outbreaker.config}}}

\item{loglike}{a set of log-likelihood functions as returned by \code{\link{outbreaker.create.loglike}}}

\item{priors}{a set of prior functions as returned by \code{\link{outbreaker.create.priors}}}

\item{posteriors}{a set of posterior functions as returned by \code{\link{outbreaker.create.posteriors}}}

\item{moves}{a set of movement functions stored in a named list as returned by \code{\link{outbreaker.create.moves}}}
}
\description{
Outbreaker2: disease outbreak reconstruction using epidemiological and genetic data
}
\examples{

## get data
data(fakeOutbreak)
dat <- fakeOutbreak$dat
w <- fakeOutbreak$w

\dontrun{
## run outbreaker
out <- outbreaker(data=list(dna=dat$dna, dates=dat$onset, w.dens=w),
config=list(n.iter=2e4, sample.every=200))
plot(out)
as.data.frame(out)

## run outbreaker, no DNA sequences
out2 <- outbreaker(data=list(dates=dat$onset, w.dens=w),
config=list(n.iter=2e4, sample.every=200))
plot(out2)
as.data.frame(out2)

}
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}
\references{
Jombart T, Cori A, Didelot X, Cauchemez S, Fraser C and Ferguson
N (2014).  Bayesian reconstruction of disease outbreaks by combining
epidemiologic and genomic data. PLoS Computational Biology.
}
\seealso{
\code{\link{outbreaker.data}} to process input data, and \code{\link{outbreaker.config}} to process/set up parameters

\code{\link{outbreaker.config}} to see default parameters / set parameters and \code{\link{outbreaker.data}} to process input data
}

