% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outbreaker.R
\name{outbreaker}
\alias{outbreaker}
\title{Outbreaker2: disease outbreak reconstruction using epidemiological and genetic data}
\usage{
outbreaker(dates, dna = NULL, w.dens, f.dens = w.dens,
  config = outbreaker.config())
}
\arguments{
\item{dates}{a vector indicating the collection dates, provided either as
integer numbers or in a usual date format such as \code{Date} or
\code{POSIXct} format. By convention, zero will indicate the oldest date.}

\item{dna}{the DNA sequences in \code{DNAbin} format (see
\code{\link[ape]{read.dna}} in the ape package); this can be imported from a
fasta file (extension .fa, .fas, or .fasta) using \code{adegenet}'s function
\link[adegenet]{fasta2DNAbin}.}

\item{w.dens}{a vector of numeric values indicating the generation time
distribution, reflecting the infectious potential of a case t=0, 1, 2, ...
time steps after infection. By convention, w.dens[1]=0, meaning that an
newly infected patient cannot be instantaneously infectious. If not
standardized, this distribution is rescaled to sum to 1.}

\item{f.dens}{similar to \code{w.dens}, except that this is the distribution
of the colonization time, i.e. time interval during which the pathogen can
be sampled from the patient.}

\item{config}{a set of settings as returned by \code{\link{outbreaker.config}}}
}
\description{
Outbreaker2: disease outbreak reconstruction using epidemiological and genetic data
}
\author{
Thibaut Jombart (\email{t.jombart@imperial.ac.uk})
}
\references{
Jombart T, Cori A, Didelot X, Cauchemez S, Fraser C and Ferguson
N (2014).  Bayesian reconstruction of disease outbreaks by combining
epidemiologic and genomic data. PLoS Computational Biology.
}
\seealso{
\code{outbreaker.data} to process input data, and \code{outbreaker.config} to process/set up parameters

\code{\link{outbreaker.config}} to see default parameters / set parameters and \code{\link{outbreaker.data}} to process input data
}

