% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.methods.R
\name{print.create_config}
\alias{print.create_config}
\title{Basic methods for processing outbreaker config}
\usage{
\method{print}{create_config}(x, ...)
}
\arguments{
\item{x}{an \code{create_config} object as returned by \code{create_config}.}

\item{...}{further arguments to be passed to other methods}
}
\description{
Several methods are defined for instances of the class \code{create_config}, returned by \code{\link{create_config}}, including: \code{print}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail_com})
}

