% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker_init_mcmc.R
\name{outbreaker_init_mcmc}
\alias{outbreaker_init_mcmc}
\title{Initialise MCMC storage for outbreaker}
\usage{
outbreaker_init_mcmc(data, param_current, param_store, loglike, priors, config)
}
\arguments{
\item{data}{A list of data items as returned by \code{\link{outbreaker_data}}.}

\item{param_current}{A list of current parameter states as returned by
\code{\link{create_param}}.}

\item{param_store}{A list of stored parameter states for MCMC output.}

\item{loglike}{A list of log-likelihood components as returned by
\code{\link{custom_likelihoods}}.}

\item{priors}{A list of prior functions as returned by
\code{\link{custom_priors}}.}

\item{config}{A list of configuration settings as returned by
\code{\link{create_config}}.}
}
\value{
The \code{param_store} list, with \code{like[1]}, \code{prior[1]},
  and \code{post[1]} filled.
}
\description{
Computes the initial log-likelihood, log-prior, and log-posterior for the
starting parameter state and writes them into \code{param_store}.
}
\seealso{
\code{\link{outbreaker}}, \code{\link{create_param}}.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com}).
}
