% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.init.mcmc.R
\name{outbreaker_init_mcmc}
\alias{outbreaker_init_mcmc}
\title{Initianilizes outputs for outbreaker}
\usage{
outbreaker_init_mcmc(param_current, param_store, loglike, priors)
}
\arguments{
\item{loglike}{a list of loglikelihood functions with enclosed data as returned by \code{create_loglike}}

\item{priors}{a list of prior functions with enclosed parameters as returned by \code{create_priors}}

\item{param}{a list of data items as returned by \code{outbreaker_create_mcmc}}
}
\description{
This function creates initial outputs and parameter states for outbreaker.
}
\author{
Thibaut Jombart (\email{t_jombart@imperial_ac_uk})
}

