% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.mcmc.store.R
\name{outbreaker_mcmc_store}
\alias{outbreaker_mcmc_store}
\title{Stores MCMC samples for outbreaker}
\usage{
outbreaker_mcmc_store(param_current, param_store, densities, step)
}
\arguments{
\item{densities}{a list containing lists of functions computing densities, named: 'loglike' (log-likelihoods), 'priors' and 'posteriors'}

\item{step}{an integer indicating the MCMC iteration being stored}

\item{param}{a list of output items as returned by \code{outbreaker_create_mcmc}}
}
\description{
This function creates stores MCMC samples for outbreaker: augmented data and parameter states, likelihood, priors and posterior.
}
\author{
Thibaut Jombart (\email{t_jombart@imperial_ac_uk})
}

