% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.move.R
\name{outbreaker_move}
\alias{outbreaker_move}
\title{Movements of augmented data and parameters for outbreaker2}
\usage{
outbreaker_move(moves, data, param_current, param_store, config, densities)
}
\arguments{
\item{moves}{a list of movement functions as returned by \code{create_moves}
(internal function)}

\item{data}{a list of data items as returned by \code{outbreaker_data}}

\item{config}{a list of settings as returned by \code{outbreaker_config}}

\item{densities}{a list containing lists of functions computing densities,
named: 'loglike' (log-likelihoods), 'priors' and 'posteriors'}

\item{param}{a list of parameters as returned by
\code{outbreaker_create_mcmc}}
}
\value{
a potentially modified list of parameters as returned by
    \code{outbreaker_create_mcmc}
}
\description{
This function moves all parameters for outbreaker2.
}
\author{
Thibaut Jombart \email{t_jombart@imperial_ac_uk}
}

