% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{chains.pal}
\alias{chains.pal}
\alias{outbreaker.palettes}
\title{Color palettes used in outbreaker}
\usage{
chains.pal(n)
}
\description{
These functions are different color palettes (color-generating functions) used in outbreaker.
}
\examples{
plot(1:8, col=chains.pal(8), cex=10, pch=20)
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

