% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{post.genetic}
\alias{post.all}
\alias{post.genetic}
\title{Posteriors of outbreaker2}
\usage{
post.genetic(D, gen.length, ances, mu)

post.all(data = data, chain = chain)
}
\arguments{
\item{D}{a matrix of pairwise genetic distances}

\item{gen.length}{length of the genetic sequences}

\item{ances}{a vector of indices of ancestors}

\item{mu}{a mutation rate}

\item{data}{a list of data items as returned by \code{outbreaker.data}}

\item{chain}{a list of output items as returned by \code{outbreaker.mcmc.init}}

\item{t.inf}{a vector of integers indicating dates of infections of the cases}

\item{log.w}{a vector of log probabilities of time intervals (between infections), starting at p(T=0)}

\item{log.f}{a vector of log probabilities of time intervals (from infection to collection), starting at p(T=0)}

\item{sampling.times}{a vector of integers indicating dates of sampling/observation/reporting of the cases}
}
\description{
These functions compute various posteriors used in outbreaker.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

