% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/priors.R
\name{prior.mu}
\alias{prior.all}
\alias{prior.mu}
\title{Priors of outbreaker2}
\usage{
prior.mu(mu)

prior.all(mu)
}
\arguments{
\item{mu}{a mutation rate}
}
\description{
These functions compute various priors used in outbreaker.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

