% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior.mu}
\alias{prior.all}
\alias{prior.mu}
\alias{prior.pi}
\title{Priors of outbreaker2}
\usage{
prior.mu(param)

prior.pi(param)

prior.all(param)
}
\arguments{
\item{param}{a list containing parameters as returned by \code{outbreaker.mcmc.init}}
}
\description{
These functions compute various priors used in outbreaker.
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

