% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ctd.R
\name{sim_ctd}
\alias{sim_ctd}
\title{Simulate contact data from a transmission tree}
\usage{
sim_ctd(ttree, eps, lambda)
}
\arguments{
\item{ttree}{A data frame or matrix with two columns, each row giving the IDs
(numeric or character) of a transmission pair.}

\item{eps}{Contact reporting coverage: probability of reporting a contact
(whether or not the pair is a transmission pair).}

\item{lambda}{Non-infectious contact rate: probability of contact between a
non-transmission pair.}
}
\value{
A data frame with columns \code{i} and \code{j} containing simulated
  contact pairs.
}
\description{
Simulates contact data from a transmission tree. The model assumes that all
transmission pairs may be observed as contacts and that there is no
false-positive reporting of contacts. The probability of contact occurring
between a non-transmission pair is given by \code{lambda}. The probability of
reporting a contact (transmission pair or not) is given by \code{eps}.
}
\examples{

## load data
x <- fake_outbreak
ttree <- data.frame(from = x$ances, to = seq_along(x$ances))

## simulate contact data with coverage of 80\% and 10\% probability of
## contact between non-transmission pairs
ctd <- outbreaker2:::sim_ctd(ttree, eps = 0.8, lambda = 0.1)

## inspect contact data
head(ctd)
}
\author{
Finlay Campbell (\email{finlaycampbell93@gmail.com}).
}
\keyword{internal}
