% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ctd.R
\name{sim_ctd}
\alias{sim_ctd}
\title{Simulate contact tracing data from a transmission tree.}
\usage{
sim_ctd(tTree, eps, lambda)
}
\arguments{
\item{tTree}{a dataframe or matrix of two columns, with each row providing
the ids (numerical or as characters) of a transmission pairs}

\item{eps}{the contact reporting coverage, defined as the probability of
reporting a contact between a transmission pair.}

\item{lambda}{the non-transmission contact rate, defined as the probability
of contact between a non-transmission pair.}
}
\description{
This function simulates contact tracing data from a transmission tree. This
model assumes that all transmission pairs have experienced contact, and that
there is no false-positive reporting of contacts. The probability of contact
occuring between a non-transmision pair is given by the parameter lambda. The
probability of reporting a contact (transmission pair or not) is given by the
parameters eps.
}
\author{
Finlay Campbell (\email{f.campbell15@imperial.ac.uk})
}
