% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_CTD.R
\name{sim_CTD}
\alias{sim_CTD}
\title{Simulate contact tracing data from a simOutbreak object}
\usage{
sim_CTD(tTree, eps, lambda)
}
\arguments{
\item{tTree}{a dataframe or matrix of two columns, with each row providing
the ids (numerical or as characters) of a transmission pairs}

\item{eps}{the contact reporting coverage, defined as the probability of
reporting a contact between a transmission pair.}

\item{lambda}{the non-transmission contact rate, defined as the probability
of contact between a non-transmisison pair.}
}
\description{
This function simulates contact tracing data from a transmission tree. The
probability of a contact occuring between a transmision pair is assumed to be
1. The probability of contact occuring between a non-transmision pair is
given by the parameter lambda. The probability of reporting a contact
(transmission pair or not) is given by the parameters eps.
}
\author{
Finlay Campbell (\email{f.campbell15@imperial.ac.uk})
}
