% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCTD.R
\name{sim_CTD}
\alias{sim_CTD}
\title{Simulate contact tracing data from a simOutbreak object}
\usage{
sim_CTD(outbreak, eps, lambda)
}
\arguments{
\item{outbreak}{a \code{simOutbreak} object}

\item{eps}{the contact reporting coverage, defined as the probability of
reporting a contact between a transmission pair.}

\item{lambda}{the non-transmission contact rate, defined as the probability
of contact between a non-transmisison pair.}
}
\description{
This function takes a simOutbreak object and returns a dataframe of contact pairs. The probabilites of reporting infectious and non-infectious contacts are provided as parameters.
}
\author{
Finlay Campbell (\email{f.campbell15@imperial.ac.uk})
}

